Eliminate SurvivalPlayer and CreativePlayer to retain data structure of initial VanillaPlayer parent Signed-off-by: Windwaker <wjcrouse913@aim.com>
