+++ src/main/java/org/spout/vanilla/material/item/Shears.java
@@ -31,8 +31,9 @@ import org.spout.api.entity.Entity;
 import org.spout.api.inventory.ItemStack;
 import org.spout.api.material.Material;
 
+import org.spout.vanilla.VanillaEventListener;
 import org.spout.vanilla.controller.living.creature.passive.Sheep;
-import org.spout.vanilla.controller.living.player.SurvivalPlayer;
+import org.spout.vanilla.controller.living.player.VanillaPlayer;
 import org.spout.vanilla.controller.object.moving.Item;
 import org.spout.vanilla.material.generic.GenericTool;
 
@@ -62,7 +63,7 @@ public class Shears extends GenericTool {
 		other.getWorld().createAndSpawnEntity(other.getPosition(), new Item(new ItemStack(Material.get((short) 35), col, rand.nextInt(3) + 1), other.getPosition().normalize()));
 
 		ItemStack holding = entity.getInventory().getCurrentItem();
-		if (entity.getController() instanceof SurvivalPlayer) {
+		if (entity.getController() instanceof VanillaPlayer && ((VanillaPlayer) entity.getController()).isSurvival()) {
 			holding.setData((short) (holding.getData() + 1));
 			entity.getInventory().setItem(holding, entity.getInventory().getCurrentSlot());
 		}
